Update to support bazel and buck tests (including plugins)

+++ jenkins/gerrit-buck-build-its-plugin.sh
@@ -7,6 +7,7 @@
 # out the proper version of Gerrit, removing the plugin if it
 # exists, then use git read-tree to put the plugin we're
 # building in place.
+
 git checkout gerrit/{branch}
 rm -rf plugins/its-{name}
 rm -rf plugins/its-base
@@ -15,20 +16,23 @@ git read-tree -u --prefix=plugins/its-base base/{branch}
 
 rm -Rf buck-out
 
-SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
-. set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
+if [ -f 'plugins/its-{name}/BUCK' ]
+then
+  SOURCE_LEVEL=$(grep "source_level" .buckconfig || echo "source_level=7")
+  . set-java.sh $(echo $SOURCE_LEVEL | cut -d '=' -f 2 | tr -d '[[:space:]]')
 
-buck build -v 3 plugins/its-{name}
+  buck build -v 3 plugins/its-{name}
 
-# Remove duplicate entries
-PLUGIN_JAR=$(ls $(pwd)/buck-out/gen/plugins/its-{name}/its-{name}*.jar)
-mkdir jar-out && pushd jar-out
-jar xf $PLUGIN_JAR && jar cmf META-INF/MANIFEST.MF $PLUGIN_JAR .
-popd
+  # Remove duplicate entries
+  PLUGIN_JAR=$(ls $(pwd)/buck-out/gen/plugins/its-{name}/its-{name}*.jar)
+  mkdir jar-out && pushd jar-out
+  jar xf $PLUGIN_JAR && jar cmf META-INF/MANIFEST.MF $PLUGIN_JAR .
+  popd
 
-# Extract version information
-PLUGIN_VERSION=$(git describe --always origin/{branch})
-echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
-jar ufm $PLUGIN_JAR MANIFEST.MF && rm MANIFEST.MF
+  # Extract version information
+  PLUGIN_VERSION=$(git describe --always origin/{branch})
+  echo -e "Implementation-Version: $PLUGIN_VERSION" > MANIFEST.MF
+  jar ufm $PLUGIN_JAR MANIFEST.MF && rm MANIFEST.MF
 
-echo "$PLUGIN_VERSION" > $PLUGIN_JAR-version
+  echo "$PLUGIN_VERSION" > $PLUGIN_JAR-version
+fi
